+++ harfrust/src/hb/aat/layout.rs
@@ -4,8 +4,7 @@ use super::map;
 use super::{layout_kerx_table, layout_morx_table, layout_trak_table};
 use crate::hb::aat::layout_common::{AatApplyContext, HB_BUFFER_SCRATCH_FLAG_AAT_HAS_DELETED};
 use crate::hb::{
-    buffer::hb_buffer_t, face::Scale, hb_font_t, hb_tag_t, ot_shape_plan::hb_ot_shape_plan_t,
-    GlyphInfo,
+    buffer::Buffer, face::Scale, hb_font_t, hb_tag_t, ot_shape_plan::hb_ot_shape_plan_t, GlyphInfo,
 };
 use crate::Feature;
 
@@ -503,7 +502,7 @@ pub const DELETED_GLYPH: u32 = 0xFFFF;
 pub fn substitute(
     plan: &hb_ot_shape_plan_t,
     face: &hb_font_t,
-    buffer: &mut hb_buffer_t,
+    buffer: &mut Buffer,
     features: &[Feature],
 ) {
     let mut aat_map = map::AatMap::default();
@@ -534,7 +533,7 @@ fn is_deleted_glyph(info: &GlyphInfo) -> bool {
 ///
 /// See <https://github.com/harfbuzz/harfbuzz/blob/2c22a65f0cb99544c36580b9703a43b5dc97a9e1/src/hb-aat-layout.cc#L337>
 #[doc(alias = "hb_aat_layout_remove_deleted_glyphs")]
-pub fn remove_deleted_glyphs(buffer: &mut hb_buffer_t) {
+pub fn remove_deleted_glyphs(buffer: &mut Buffer) {
     if (buffer.scratch_flags & HB_BUFFER_SCRATCH_FLAG_AAT_HAS_DELETED) != 0 {
         buffer.delete_glyphs_inplace(is_deleted_glyph);
     }
@@ -544,12 +543,7 @@ pub fn remove_deleted_glyphs(buffer: &mut hb_buffer_t) {
 ///
 /// See <https://github.com/harfbuzz/harfbuzz/blob/2c22a65f0cb99544c36580b9703a43b5dc97a9e1/src/hb-aat-layout.cc#L363>
 #[doc(alias = "hb_aat_layout_position")]
-pub fn position(
-    plan: &hb_ot_shape_plan_t,
-    face: &hb_font_t,
-    scale: Scale,
-    buffer: &mut hb_buffer_t,
-) {
+pub fn position(plan: &hb_ot_shape_plan_t, face: &hb_font_t, scale: Scale, buffer: &mut Buffer) {
     let mut c = AatApplyContext::new(plan, face, scale, buffer);
     layout_kerx_table::apply(&mut c);
 }
@@ -563,7 +557,7 @@ pub fn track(
     face: &hb_font_t,
     scale: Scale,
     point_size: Option<f32>,
-    buffer: &mut hb_buffer_t,
+    buffer: &mut Buffer,
 ) {
     layout_trak_table::apply(plan, face, scale, point_size, buffer);
 }