+++ harfrust/src/hb/ot_shaper_indic.rs
@@ -576,7 +576,7 @@ fn override_features(planner: &mut hb_ot_shape_planner_t) {
     planner.ot_map.add_gsub_pause(Some(syllabic_clear_var)); // Don't need syllables anymore.
 }
 
-fn preprocess_text(_: &hb_ot_shape_plan_t, _: &mut FontFuncsDispatch, buffer: &mut hb_buffer_t) {
+fn preprocess_text(_: &hb_ot_shape_plan_t, _: &mut FontFuncsDispatch, buffer: &mut Buffer) {
     super::ot_shaper_vowel_constraints::preprocess_text_vowel_constraints(buffer);
 }
 
@@ -608,7 +608,7 @@ fn compose(_: &hb_ot_shape_normalize_context_t, a: Codepoint, b: Codepoint) -> O
     crate::unicode::compose(a, b)
 }
 
-fn setup_masks(_: &hb_ot_shape_plan_t, _: &mut FontFuncsDispatch, buffer: &mut hb_buffer_t) {
+fn setup_masks(_: &hb_ot_shape_plan_t, _: &mut FontFuncsDispatch, buffer: &mut Buffer) {
     buffer.allocate_var(GlyphInfo::INDIC_CATEGORY_VAR);
     buffer.allocate_var(GlyphInfo::INDIC_POSITION_VAR);
 
@@ -619,11 +619,7 @@ fn setup_masks(_: &hb_ot_shape_plan_t, _: &mut FontFuncsDispatch, buffer: &mut h
     }
 }
 
-fn setup_syllables(
-    _: &hb_ot_shape_plan_t,
-    _: &mut FontFuncsDispatch,
-    buffer: &mut hb_buffer_t,
-) -> bool {
+fn setup_syllables(_: &hb_ot_shape_plan_t, _: &mut FontFuncsDispatch, buffer: &mut Buffer) -> bool {
     buffer.allocate_var(GlyphInfo::SYLLABLE_VAR);
 
     super::ot_shaper_indic_machine::find_syllables_indic(buffer);
@@ -642,7 +638,7 @@ fn setup_syllables(
 fn initial_reordering(
     plan: &hb_ot_shape_plan_t,
     font_funcs: &mut FontFuncsDispatch,
-    buffer: &mut hb_buffer_t,
+    buffer: &mut Buffer,
 ) -> bool {
     use super::ot_shaper_indic_machine::SyllableType;
 
@@ -677,7 +673,7 @@ fn update_consonant_positions(
     plan: &hb_ot_shape_plan_t,
     indic_plan: &IndicShapePlan,
     font_funcs: &mut FontFuncsDispatch,
-    buffer: &mut hb_buffer_t,
+    buffer: &mut Buffer,
 ) {
     let mut virama_glyph = None;
     if indic_plan.config.virama != 0 {
@@ -763,7 +759,7 @@ fn initial_reordering_syllable(
     font_funcs: &mut FontFuncsDispatch,
     start: usize,
     end: usize,
-    buffer: &mut hb_buffer_t,
+    buffer: &mut Buffer,
 ) {
     use super::ot_shaper_indic_machine::SyllableType;
 
@@ -798,7 +794,7 @@ fn initial_reordering_consonant_syllable(
     font_funcs: &mut FontFuncsDispatch,
     start: usize,
     end: usize,
-    buffer: &mut hb_buffer_t,
+    buffer: &mut Buffer,
 ) {
     let face = font_funcs.font();
     // https://github.com/harfbuzz/harfbuzz/issues/435#issuecomment-335560167
@@ -1322,7 +1318,7 @@ fn initial_reordering_standalone_cluster(
     face: &mut FontFuncsDispatch,
     start: usize,
     end: usize,
-    buffer: &mut hb_buffer_t,
+    buffer: &mut Buffer,
 ) {
     // We treat placeholder/dotted-circle as if they are consonants, so we
     // should just chain.  Only if not in compatibility mode that is...
@@ -1332,7 +1328,7 @@ fn initial_reordering_standalone_cluster(
 fn final_reordering(
     plan: &hb_ot_shape_plan_t,
     face: &mut FontFuncsDispatch,
-    buffer: &mut hb_buffer_t,
+    buffer: &mut Buffer,
 ) -> bool {
     if buffer.is_empty() {
         return false;
@@ -1353,7 +1349,7 @@ fn final_reordering_impl(
     face: &mut FontFuncsDispatch,
     start: usize,
     end: usize,
-    buffer: &mut hb_buffer_t,
+    buffer: &mut Buffer,
 ) {
     let indic_plan = plan.data::<IndicShapePlan>();